// packages/client/src/store/modal/types.ts
export enum ModalType {
  SignIn = 'signin',
  SignUp = 'signup',
  RestorePassword = 'RestorePassword',
  ChatRain = 'chatRain',
  ChatTipping = 'chatTipping',
  OAuthSignup = 'oauthSignup',
  SignUpMetamask = 'signupMetamask',
  SignUpTelegram = 'signupTelegram',
  SelectDate = 'selectDate',
  WithdrawalConfirm = 'withdrawalConfirm',
  WithdrawalSuccess = 'withdrawalSuccess',
  SportBetsDate = 'sportBetsDate',
  SportBetsFilter = 'sportBetsFilter',
  TransactionFilter = 'transactionFilter',
  GameLiveStats = 'gameLiveStats',
}